ADMIN/Services/delete_all_mongo.py
from pymongo import MongoClient
import os
from dotenv import load_dotenv

load_dotenv()

mongo_client = MongoClient(os.getenv("MONGODB_URI"))
db = mongo_client["research_ai_database"]
collection = db["pdfs"]

collection.delete_many({})